Made "this" be an outer scope value for each method
Previously we relied on an invisible "this" parameter on each methods
for instance access. This created issues when saving into a variable a
method of an inline instance.
This new approach creates an artifical capture scope with a "this" local
variable and replaces the class' methods capture for this new one.
Note that this approach is flawed when declaring a non-global class
(which should be prevented altogether)